feat: Add TheMovieDB json parsing and movie poster display
Added json parsing and image pulling using the volley library from AOSP. This was done because it offers a very nice way of managing the HTTP requests, JSON parsing, and Image loading without a lot of the needed boilerplate using the normal method. The gui itself shows up correctly, though there is no handling of clicking on the images in order to open the movie detail screen, which has still yet to be coded.