feat: Add Movie detail view from clicking on thumbnail
Added a detailed overview of the movie information that was pulled from theMovieDB for the movies, accessed by tapping on the poster image. It shows the poster image, movie title, release date, average rating, and synopsis for the movie in question. Also a slight bit of refactoring of string literals into the string resource file.