Convert project to use the Retrofit and Picasso libraries rather than Volley
Decided to convert the project to using the Retrofit and Picasso libraries and remove volley. Reasoning for this is that the documentation existing for Volley is practically nonexistant and Retrofit is regarded at least in some circles as being a better library overall to volley. This is mostly a personal preference, but decided to try going with the popular option just to see how it works.