Add Review and Trailer information to the Movie Detail View
The trailer and review information from a given movie can now be accessed from it's detail page. The trailers, when clicked, will start an intent to try to view the movie within either Youtube or chrome, depending on what the user has installed onto their device and what their default settings are.