Add new source endpoints and additional data to the database
Change the syncing to also support pulling from the Upcoming movie list and the Now Playing source list. They are not reflected in the UI just yet, as there will be a large overhaul of the main grid ui to make use of the design library, and the source list will be placed in a drawer. Also added additional data to the database for if we need it, as we pull the data down anyway, and having it in the database will be beneficial for the Movie Detail fragment redesign.