Commit forgotten files and add Workaround for Scroll state not getting restored
Forgot to add the icons for the tabs and header background, so added them to the git repository. Also after much investigation committed a very hacky way of ensuring that the state is restored correctly on close and rotation. Essentially the GridLayoutManager attempts to restore the state before the RecyclerView is able to get the data it needs from the cursor. This means that it attempts to restore the scrolling state on an empty RecyclerView, and so when it goes to scroll to it's old position it is unable to. The hack is essentially to force the LayoutManager to wait for 250ms while the data gets loaded, then attempt to restore the state. I will continue to look for alternatives to this as I am not satisfied with the solution.