Fix configuration change not restoring state correctly and remove hack
For the configuration change it was discovered that because there was no default image for the poster view items, it was incorrectly calculating the scroll amount required. The scroll state is still reset on moving to a different tab, but this is preferable to using a delay hack in order to set the position.